Refactored mod integration to correct packages
