+++ src/main/java/crazypants/enderio/enchantment/EnchantmentSoulBound.java
@@ -9,8 +9,8 @@ import com.enderio.core.api.common.enchant.IAdvancedEnchant;
 import crazypants.enderio.EnderIO;
 import crazypants.enderio.Log;
 import crazypants.enderio.config.Config;
-import crazypants.util.BaublesUtil;
-import crazypants.util.GalacticraftUtil;
+import crazypants.enderio.integration.baubles.BaublesUtil;
+import crazypants.enderio.integration.galacticraft.GalacticraftUtil;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.enchantment.EnchantmentHelper;
 import net.minecraft.enchantment.EnumEnchantmentType;
