+++ src/main/java/crazypants/enderio/integration/forestry/NaturalistEyeUpgrade.java
@@ -1,7 +1,8 @@
-package crazypants.enderio.item.darksteel.upgrade;
+package crazypants.enderio.integration.forestry;
 
 import crazypants.enderio.config.Config;
 import crazypants.enderio.item.darksteel.DarkSteelItems;
+import crazypants.enderio.item.darksteel.upgrade.AbstractUpgrade;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.inventory.EntityEquipmentSlot;
 import net.minecraft.item.Item;
