+++ src/main/java/crazypants/enderio/integration/ic2e/RubberTreeFarmerIC2exp.java
@@ -1,7 +1,8 @@
-package crazypants.enderio.machine.farm.farmers;
+package crazypants.enderio.integration.ic2e;
 
 import java.util.Random;
 
+import crazypants.enderio.machine.farm.farmers.RubberTreeFarmer;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
