+++ src/main/java/crazypants/enderio/integration/thaumcraft/ManaBeanFarmer.java
@@ -1,6 +1,7 @@
-package crazypants.enderio.machine.farm.farmers;
+package crazypants.enderio.integration.thaumcraft;
 
 import crazypants.enderio.machine.farm.TileFarmStation;
+import crazypants.enderio.machine.farm.farmers.CustomSeedFarmer;
 import net.minecraft.block.Block;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
