+++ src/main/java/crazypants/enderio/item/ItemMagnet.java
@@ -13,11 +13,11 @@ import baubles.api.IBauble;
 import crazypants.enderio.EnderIOTab;
 import crazypants.enderio.ModObject;
 import crazypants.enderio.config.Config;
+import crazypants.enderio.integration.baubles.BaublesUtil;
 import crazypants.enderio.item.darksteel.DarkSteelItems;
 import crazypants.enderio.item.darksteel.upgrade.IRenderUpgrade;
 import crazypants.enderio.power.AbstractPoweredItem;
 import crazypants.enderio.power.PowerDisplayUtil;
-import crazypants.util.BaublesUtil;
 import crazypants.util.NbtValue;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.EntityLivingBase;
