+++ src/main/java/crazypants/enderio/item/KeyTracker.java
@@ -13,6 +13,7 @@ import com.enderio.core.common.util.ChatUtil;
 import crazypants.enderio.EnderIO;
 import crazypants.enderio.api.tool.IConduitControl;
 import crazypants.enderio.conduit.ConduitDisplayMode;
+import crazypants.enderio.integration.baubles.BaublesUtil;
 import crazypants.enderio.item.MagnetController.ActiveMagnet;
 import crazypants.enderio.item.PacketMagnetState.SlotType;
 import crazypants.enderio.item.darksteel.DarkSteelController;
@@ -25,7 +26,6 @@ import crazypants.enderio.item.darksteel.upgrade.SpeedUpgrade;
 import crazypants.enderio.network.PacketHandler;
 import crazypants.enderio.sound.SoundHelper;
 import crazypants.enderio.sound.SoundRegistry;
-import crazypants.util.BaublesUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.client.multiplayer.PlayerControllerMP;
