+++ src/main/java/crazypants/enderio/item/darksteel/ItemDarkSteelArmor.java
@@ -17,16 +17,16 @@ import com.google.common.collect.Multimap;
 import crazypants.enderio.EnderIO;
 import crazypants.enderio.EnderIOTab;
 import crazypants.enderio.config.Config;
+import crazypants.enderio.integration.forestry.ApiaristArmorUpgrade;
+import crazypants.enderio.integration.forestry.NaturalistEyeUpgrade;
 import crazypants.enderio.item.IHasPlayerRenderer;
 import crazypants.enderio.item.PowerBarOverlayRenderHelper;
 import crazypants.enderio.item.darksteel.PacketUpgradeState.Type;
-import crazypants.enderio.item.darksteel.upgrade.ApiaristArmorUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.ElytraUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.EnergyUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.GliderUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.IDarkSteelUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.IRenderUpgrade;
-import crazypants.enderio.item.darksteel.upgrade.NaturalistEyeUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.NightVisionUpgrade;
 import crazypants.enderio.item.darksteel.upgrade.PaintedHelmetLayer;
 import crazypants.enderio.item.darksteel.upgrade.SoundDetectorUpgrade;
