+++ src/main/java/crazypants/enderio/machine/capbank/BlockCapBank.java
@@ -17,6 +17,7 @@ import crazypants.enderio.EnderIO;
 import crazypants.enderio.GuiID;
 import crazypants.enderio.ModObject;
 import crazypants.enderio.api.redstone.IRedstoneConnectable;
+import crazypants.enderio.integration.baubles.BaublesUtil;
 import crazypants.enderio.integration.waila.IWailaInfoProvider;
 import crazypants.enderio.machine.IoMode;
 import crazypants.enderio.machine.capbank.network.CapBankClientNetwork;
@@ -44,7 +45,6 @@ import crazypants.enderio.render.registry.SmartModelAttacher;
 import crazypants.enderio.render.registry.TextureRegistry;
 import crazypants.enderio.render.registry.TextureRegistry.TextureSupplier;
 import crazypants.enderio.tool.ToolUtil;
-import crazypants.util.BaublesUtil;
 import crazypants.util.NullHelper;
 import crazypants.util.TextUtil;
 import info.loenwind.autosave.Reader;
