+++ src/main/java/crazypants/enderio/machine/capbank/ContainerCapBank.java
@@ -7,8 +7,8 @@ import javax.annotation.Nullable;
 import com.enderio.core.common.ContainerEnder;
 import com.enderio.core.common.util.ArrayInventory;
 
+import crazypants.enderio.integration.baubles.BaublesUtil;
 import crazypants.enderio.machine.capbank.network.InventoryImpl;
-import crazypants.util.BaublesUtil;
 import crazypants.util.ShadowInventory;
 import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.player.EntityPlayer;
